requirements.txt
beautifulsoup4
requests
googlemaps
python-dotenv
lxml